Add files via upload
The program is written with C++. It requires opencv_world454d.dll. However, its size is too large for GitHub. Therefore, the program is uploaded without dll.

To test the program, /ConcurrencyTask/x64/Debug/ConcurrencyTask.exe should be run. When running, the program will require full image path  to the image that is going to be handled, and the result is saved in result.jpg.